move util functions to util.rs
